# docs/frontleaves_yggleaf_swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/game-profile/users/{user_id}/quota": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "管理员调整指定用户的游戏档案配额总额度（相对变化量）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "管理员-游戏档案接口"
                ],
                "summary": "[超管] 调整用户游戏档案配额",
                "parameters": [
                    {
                        "type": "string",
                        "description": "目标用户 ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "调整配额请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.AdminAdjustQuotaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "调整成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.GameProfileQuota"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "403": {
                        "description": "需要超级管理员权限",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "用户配额不存在",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
...
        "/game-profile": {
            "get": {
                "description": "获取当前用户的所有游戏档案列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "游戏档案接口"
                ],
                "summary": "[玩家] 获取游戏档案列表",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.GameProfileListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "根据当前登录用户创建游戏档案，UUID 由系统按 UUIDv7 自动生成",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "游戏档案接口"
                ],
                "summary": "[玩家] 创建游戏档案",
                "parameters": [
                    {
                        "description": "创建游戏档案请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.AddGameProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.GameProfileResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "资源不存在",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "409": {
                        "description": "资源冲突",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "503": {
                        "description": "资源耗尽",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
        "/game-profile/quota": {
            "get": {
                "description": "获取当前用户的游戏档案配额信息，包括总额度与已使用额度",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "游戏档案接口"
                ],
                "summary": "[玩家] 获取游戏档案配额",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/entity.GameProfileQuota"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
        "/game-profile/{profile_id}": {
            "get": {
                "description": "根据档案 ID 获取游戏档案详情，包含关联的皮肤和披风信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "游戏档案接口"
                ],
                "summary": "[玩家] 获取游戏档案详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "游戏档案 ID",
                        "name": "profile_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.GameProfileResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "资源不存在",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
        "/game-profile/{profile_id}/cape": {
            "patch": {
                "description": "统一的披风设置接口，传入 cape_library_id 为装备，传 null 或空字符串为卸下",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "游戏档案接口"
                ],
                "summary": "[玩家] 设置/卸下披风",
                "parameters": [
                    {
                        "type": "string",
                        "description": "游戏档案 ID",
                        "name": "profile_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "设置披风请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.SetCapeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.GameProfileResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "资源不存在",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
        "/game-profile/{profile_id}/skin": {
            "patch": {
                "description": "统一的皮肤设置接口，传入 skin_library_id 为装备，传 null 或空字符串为卸下",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "游戏档案接口"
                ],
                "summary": "[玩家] 设置/卸下皮肤",
                "parameters": [
                    {
                        "type": "string",
                        "description": "游戏档案 ID",
                        "name": "profile_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "设置皮肤请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.SetSkinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "操作成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.GameProfileResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "资源不存在",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
        "/game-profile/{profile_id}/username": {
            "patch": {
                "description": "根据档案 ID 修改游戏档案用户名",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "游戏档案接口"
                ],
                "summary": "[玩家] 修改用户名",
                "parameters": [
                    {
                        "type": "string",
                        "description": "游戏档案 ID",
                        "name": "profile_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "修改用户名请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.ChangeUsernameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "修改成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/xBase.BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.GameProfileResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "401": {
                        "description": "未授权",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "404": {
                        "description": "资源不存在",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    },
                    "409": {
                        "description": "资源冲突",
                        "schema": {
                            "$ref": "#/definitions/xBase.BaseResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.GameProfileQuota": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "total": {
                    "description": "总额度",
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                },
                "used": {
                    "description": "已使用额度",
                    "type": "integer"
                },
                "user": {
                    "description": "----------\n 外键约束\n----------",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.User"
                        }
                    ]
                },
                "user_id": {
                    "description": "关联用户ID",
                    "type": "integer"
                }
            }
        },
...
        "user.AddGameProfileRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "user.AdminAdjustQuotaRequest": {
            "type": "object",
            "required": [
                "delta"
            ],
            "properties": {
                "delta": {
                    "description": "配额变化量（正数增加，负数减少）",
                    "type": "integer"
                },
                "remark": {
                    "description": "备注（可选，最长 255 字符）",
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
...
        "user.GameProfileListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "description": "游戏档案列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/user.GameProfileResponse"
                    }
                }
            }
        },
...
        "user.SetCapeRequest": {
            "type": "object",
            "properties": {
                "cape_library_id": {
                    "description": "披风库 ID（null 或空字符串表示卸下）",
                    "type": "string"
                }
            }
        },
...
        "xBase.BaseResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "context": {
                    "type": "string"
                },
                "data": {},
                "error_message": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "output": {
                    "type": "string"
                },
                "overhead": {
                    "type": "integer"
                }
            }
        },
...
    }
}