# internal/handler/game_profile.go
package handler

import (
	xError "github.com/bamboo-services/bamboo-base-go/common/error"
	xSnowflake "github.com/bamboo-services/bamboo-base-go/common/snowflake"
	xUtil "github.com/bamboo-services/bamboo-base-go/common/utility"
	xResult "github.com/bamboo-services/bamboo-base-go/major/result"
	apiUser "github.com/frontleaves-mc/frontleaves-yggleaf/api/user"
	"github.com/gin-gonic/gin"
	bSdkUtil "github.com/phalanx-labs/beacon-sso-sdk/utility"
)

// AddGameProfile 创建当前用户的游戏档案
//
// @Summary     [玩家] 创建游戏档案
// @Description 根据当前登录用户创建游戏档案，UUID 由系统按 UUIDv7 自动生成
// @Tags        游戏档案接口
// @Accept      json
// @Produce     json
// @Param       request body apiUser.AddGameProfileRequest true "创建游戏档案请求"
// @Success     200   {object}  xBase.BaseResponse{data=apiUser.GameProfileResponse} "创建成功"
// @Failure     400   {object}  xBase.BaseResponse                               "请求参数错误"
// @Failure     401   {object}  xBase.BaseResponse                               "未授权"
// @Failure     404   {object}  xBase.BaseResponse                               "资源不存在"
// @Failure     409   {object}  xBase.BaseResponse                               "资源冲突"
// @Failure     503   {object}  xBase.BaseResponse                               "资源耗尽"
// @Router      /game-profile [POST]
func (h *GameProfileHandler) AddGameProfile(ctx *gin.Context) {
	h.log.Info(ctx, "AddGameProfile - 创建游戏档案")

	req := xUtil.Bind(ctx, &apiUser.AddGameProfileRequest{}).Data()
	if req == nil {
		return
	}

	userinfo, xErr := h.service.oauthLogic.Userinfo(ctx, bSdkUtil.GetAuthorization(ctx))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	userID, err := xSnowflake.ParseSnowflakeID(userinfo.Sub)
	if err != nil {
		_ = ctx.Error(xError.NewError(ctx, xError.ParameterError, "解析用户 ID 失败", true, err))
		return
	}

	profile, xErr := h.service.gameProfileLogic.AddGameProfile(ctx.Request.Context(), userID, req.Name)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.SuccessHasData(ctx, "创建游戏档案成功", gameProfileDTOToResponse(profile))
}

// ChangeUsername 修改当前用户指定游戏档案用户名
//
// @Summary     [玩家] 修改用户名
// @Description 根据档案 ID 修改游戏档案用户名
// @Tags        游戏档案接口
// @Accept      json
// @Produce     json
// @Param       profile_id path string true "游戏档案 ID"
// @Param       request body apiUser.ChangeUsernameRequest true "修改用户名请求"
// @Success     200   {object}  xBase.BaseResponse{data=apiUser.GameProfileResponse} "修改成功"
// @Failure     400   {object}  xBase.BaseResponse                               "请求参数错误"
// @Failure     401   {object}  xBase.BaseResponse                               "未授权"
// @Failure     404   {object}  xBase.BaseResponse                               "资源不存在"
// @Failure     409   {object}  xBase.BaseResponse                               "资源冲突"
// @Router      /game-profile/{profile_id}/username [PATCH]
func (h *GameProfileHandler) ChangeUsername(ctx *gin.Context) {
	h.log.Info(ctx, "ChangeUsername - 修改游戏档案用户名")

	profileID, err := xSnowflake.ParseSnowflakeID(ctx.Param("profile_id"))
	if err != nil {
		_ = ctx.Error(xError.NewError(ctx, xError.ParameterError, "解析游戏档案 ID 失败", true, err))
		return
	}

	req := xUtil.Bind(ctx, &apiUser.ChangeUsernameRequest{}).Data()
	if req == nil {
		return
	}

	userinfo, xErr := h.service.oauthLogic.Userinfo(ctx, bSdkUtil.GetAuthorization(ctx))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	userID, err := xSnowflake.ParseSnowflakeID(userinfo.Sub)
	if err != nil {
		_ = ctx.Error(xError.NewError(ctx, xError.ParameterError, "解析用户 ID 失败", true, err))
		return
	}

	updatedProfile, xErr := h.service.gameProfileLogic.ChangeUsername(ctx.Request.Context(), userID, profileID, req.NewName)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.SuccessHasData(ctx, "修改游戏档案用户名成功", gameProfileDTOToResponse(updatedProfile))
}

// GetGameProfileDetail 获取指定游戏档案详情
//
// @Summary     [玩家] 获取游戏档案详情
// @Description 根据档案 ID 获取游戏档案详情，包含关联的皮肤和披风信息
// @Tags        游戏档案接口
// @Accept      json
// @Produce     json
// @Param       profile_id path string true "游戏档案 ID"
// @Success     200   {object}  xBase.BaseResponse{data=apiUser.GameProfileResponse} "获取成功"
// @Failure     400   {object}  xBase.BaseResponse                                     "请求参数错误"
// @Failure     401   {object}  xBase.BaseResponse                                     "未授权"
// @Failure     404   {object}  xBase.BaseResponse                                     "资源不存在"
// @Router      /game-profile/{profile_id} [GET]
func (h *GameProfileHandler) GetGameProfileDetail(ctx *gin.Context) {
	h.log.Info(ctx, "GetGameProfileDetail - 获取游戏档案详情")

	profileID, err := xSnowflake.ParseSnowflakeID(ctx.Param("profile_id"))
	if err != nil {
		_ = ctx.Error(xError.NewError(ctx, xError.ParameterError, "解析游戏档案 ID 失败", true, err))
		return
	}

	userinfo, xErr := h.service.oauthLogic.Userinfo(ctx, bSdkUtil.GetAuthorization(ctx))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	userID, err := xSnowflake.ParseSnowflakeID(userinfo.Sub)
	if err != nil {
		_ = ctx.Error(xError.NewError(ctx, xError.ParameterError, "解析用户 ID 失败", true, err))
		return
	}

	profile, xErr := h.service.gameProfileLogic.GetGameProfileDetail(ctx.Request.Context(), userID, profileID)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.SuccessHasData(ctx, "获取游戏档案详情成功", gameProfileDTOToResponse(profile))
}

// ListGameProfiles 获取当前用户的游戏档案列表
//
// @Summary     [玩家] 获取游戏档案列表
// @Description 获取当前用户的所有游戏档案列表
// @Tags        游戏档案接口
// @Accept      json
// @Produce     json
// @Success     200   {object}  xBase.BaseResponse{data=apiUser.GameProfileListResponse} "获取成功"
// @Failure     400   {object}  xBase.BaseResponse                                       "请求参数错误"
// @Failure     401   {object}  xBase.BaseResponse                                       "未授权"
// @Router      /game-profile [GET]
func (h *GameProfileHandler) ListGameProfiles(ctx *gin.Context) {
	h.log.Info(ctx, "ListGameProfiles - 获取游戏档案列表")

	userinfo, xErr := h.service.oauthLogic.Userinfo(ctx, bSdkUtil.GetAuthorization(ctx))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	userID, err := xSnowflake.ParseSnowflakeID(userinfo.Sub)
	if err != nil {
		_ = ctx.Error(xError.NewError(ctx, xError.ParameterError, "解析用户 ID 失败", true, err))
		return
	}

	profiles, xErr := h.service.gameProfileLogic.ListGameProfiles(ctx.Request.Context(), userID)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	response := apiUser.GameProfileListResponse{
		Items: gameProfileDTOsToResponses(profiles),
	}

	xResult.SuccessHasData(ctx, "获取游戏档案列表成功", response)
}

// GetQuota 获取当前用户的游戏档案配额
//
// @Summary     [玩家] 获取游戏档案配额
// @Description 获取当前用户的游戏档案配额信息，包括总额度与已使用额度
// @Tags        游戏档案接口
// @Accept      json
// @Produce     json
// @Success     200   {object}  xBase.BaseResponse{data=entity.GameProfileQuota} "获取成功"
// @Failure     400   {object}  xBase.BaseResponse                               "请求参数错误"
// @Failure     401   {object}  xBase.BaseResponse                               "未授权"
// @Router      /game-profile/quota [GET]
func (h *GameProfileHandler) GetQuota(ctx *gin.Context) {
	h.log.Info(ctx, "GetQuota - 获取游戏档案配额")

	userinfo, xErr := h.service.oauthLogic.Userinfo(ctx, bSdkUtil.GetAuthorization(ctx))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	userID, err := xSnowflake.ParseSnowflakeID(userinfo.Sub)
	if err != nil {
		_ = ctx.Error(xError.NewError(ctx, xError.ParameterError, "解析用户 ID 失败", true, err))
		return
	}

	quota, xErr := h.service.gameProfileLogic.GetQuota(ctx.Request.Context(), userID)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.SuccessHasData(ctx, "获取游戏档案配额成功", quota)
}

// AdjustQuotaAdmin 管理员调整指定用户的游戏档案配额
//
// @Summary     [超管] 调整用户游戏档案配额
// @Description 管理员调整指定用户的游戏档案配额总额度（相对变化量）
// @Tags        管理员-游戏档案接口
// @Accept      json
// @Produce     json
// @Param       user_id path string true "目标用户 ID"
// @Param       request body apiUser.AdminAdjustQuotaRequest true "调整配额请求"
// @Success     200 {object} xBase.BaseResponse{data=entity.GameProfileQuota} "调整成功"
// @Failure     400 {object} xBase.BaseResponse "请求参数错误"
// @Failure     401 {object} xBase.BaseResponse "未授权"
// @Failure     403 {object} xBase.BaseResponse "需要超级管理员权限"
// @Failure     404 {object} xBase.BaseResponse "用户配额不存在"
// @Security    BearerAuth
// @Router      /admin/game-profile/users/{user_id}/quota [POST]
func (h *GameProfileHandler) AdjustQuotaAdmin(ctx *gin.Context) {
	h.log.Info(ctx, "AdjustQuotaAdmin - 管理员调整游戏档案配额")

	targetUserID, err := xSnowflake.ParseSnowflakeID(ctx.Param("user_id"))
	if err != nil {
		_ = ctx.Error(xError.NewError(ctx, xError.ParameterError, "解析目标用户 ID 失败", true, err))
		return
	}

	req := xUtil.Bind(ctx, &apiUser.AdminAdjustQuotaRequest{}).Data()
	if req == nil {
		return
	}

	userinfo, xErr := h.service.oauthLogic.Userinfo(ctx, bSdkUtil.GetAuthorization(ctx))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	operatorID, err := xSnowflake.ParseSnowflakeID(userinfo.Sub)
	if err != nil {
		_ = ctx.Error(xError.NewError(ctx, xError.ParameterError, "解析操作者 ID 失败", true, err))
		return
	}

	quota, xErr := h.service.gameProfileLogic.AdjustQuotaAdmin(ctx.Request.Context(), operatorID, targetUserID, req.Delta, req.Remark)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.SuccessHasData(ctx, "调整游戏档案配额成功", quota)
}

// ==================== Set Handlers (Unified Equip/Unequip) ====================

// SetSkin 统一设置/卸下游戏档案皮肤
//
// @Summary     [玩家] 设置/卸下皮肤
// @Description 统一的皮肤设置接口，传入 skin_library_id 为装备，传 null 或空字符串为卸下
// @Tags        游戏档案接口
// @Accept      json
// @Produce     json
// @Param       profile_id path string true "游戏档案 ID"
// @Param       request body apiUser.SetSkinRequest true "设置皮肤请求"
// @Success     200 {object} xBase.BaseResponse{data=apiUser.GameProfileResponse} "操作成功"
// @Failure     400 {object} xBase.BaseResponse "请求参数错误"
// @Failure     401 {object} xBase.BaseResponse "未授权"
// @Failure     404 {object} xBase.BaseResponse "资源不存在"
// @Router      /game-profile/{profile_id}/skin [PATCH]
func (h *GameProfileHandler) SetSkin(ctx *gin.Context) {
	h.log.Info(ctx, "SetSkin - 设置/卸下皮肤")

	profileID, err := xSnowflake.ParseSnowflakeID(ctx.Param("profile_id"))
	if err != nil {
		_ = ctx.Error(xError.NewError(ctx, xError.ParameterError, "解析游戏档案 ID 失败", true, err))
		return
	}

	req := xUtil.Bind(ctx, &apiUser.SetSkinRequest{}).Data()
	if req == nil {
		return
	}

	userinfo, xErr := h.service.oauthLogic.Userinfo(ctx, bSdkUtil.GetAuthorization(ctx))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	userID, err := xSnowflake.ParseSnowflakeID(userinfo.Sub)
	if err != nil {
		_ = ctx.Error(xError.NewError(ctx, xError.ParameterError, "解析用户 ID 失败", true, err))
		return
	}

	var skinLibraryID *xSnowflake.SnowflakeID
	if req.SkinLibraryID != nil && *req.SkinLibraryID != "" {
		parsed, parseErr := xSnowflake.ParseSnowflakeID(*req.SkinLibraryID)
		if parseErr != nil {
			_ = ctx.Error(xError.NewError(ctx, xError.ParameterError, "解析皮肤库 ID 失败", true, parseErr))
			return
		}
		skinLibraryID = &parsed
	}

	profile, xErr := h.service.gameProfileLogic.SetSkin(ctx.Request.Context(), userID, profileID, skinLibraryID)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.SuccessHasData(ctx, "设置皮肤成功", gameProfileDTOToResponse(profile))
}

// SetCape 统一设置/卸下游戏档案披风
//
// @Summary     [玩家] 设置/卸下披风
// @Description 统一的披风设置接口，传入 cape_library_id 为装备，传 null 或空字符串为卸下
// @Tags        游戏档案接口
// @Accept      json
// @Produce     json
// @Param       profile_id path string true "游戏档案 ID"
// @Param       request body apiUser.SetCapeRequest true "设置披风请求"
// @Success     200 {object} xBase.BaseResponse{data=apiUser.GameProfileResponse} "操作成功"
// @Failure     400 {object} xBase.BaseResponse "请求参数错误"
// @Failure     401 {object} xBase.BaseResponse "未授权"
// @Failure     404 {object} xBase.BaseResponse "资源不存在"
// @Router      /game-profile/{profile_id}/cape [PATCH]
func (h *GameProfileHandler) SetCape(ctx *gin.Context) {
	h.log.Info(ctx, "SetCape - 设置/卸下披风")

	profileID, err := xSnowflake.ParseSnowflakeID(ctx.Param("profile_id"))
	if err != nil {
		_ = ctx.Error(xError.NewError(ctx, xError.ParameterError, "解析游戏档案 ID 失败", true, err))
		return
	}

	req := xUtil.Bind(ctx, &apiUser.SetCapeRequest{}).Data()
	if req == nil {
		return
	}

	userinfo, xErr := h.service.oauthLogic.Userinfo(ctx, bSdkUtil.GetAuthorization(ctx))
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	userID, err := xSnowflake.ParseSnowflakeID(userinfo.Sub)
	if err != nil {
		_ = ctx.Error(xError.NewError(ctx, xError.ParameterError, "解析用户 ID 失败", true, err))
		return
	}

	var capeLibraryID *xSnowflake.SnowflakeID
	if req.CapeLibraryID != nil && *req.CapeLibraryID != "" {
		parsed, parseErr := xSnowflake.ParseSnowflakeID(*req.CapeLibraryID)
		if parseErr != nil {
			_ = ctx.Error(xError.NewError(ctx, xError.ParameterError, "解析披风库 ID 失败", true, parseErr))
			return
		}
		capeLibraryID = &parsed
	}

	profile, xErr := h.service.gameProfileLogic.SetCape(ctx.Request.Context(), userID, profileID, capeLibraryID)
	if xErr != nil {
		_ = ctx.Error(xErr)
		return
	}

	xResult.SuccessHasData(ctx, "设置披风成功", gameProfileDTOToResponse(profile))
}